refactor(habitat): enforce router import shape
