+++ tools/habitat-harness/src/service/modules/verify/module.ts
@@ -1,28 +1,128 @@
 import type { GitProviderService } from "@internal/habitat-harness/providers/git/index";
 import type { GraphiteProviderService } from "@internal/habitat-harness/providers/graphite/index";
 import type { NxProviderService } from "@internal/habitat-harness/providers/nx/index";
+import { spawnResultFromCommandResult } from "@internal/habitat-harness/resources/command/index";
 import type { epochMillisToIsoString } from "@internal/habitat-harness/resources/platform/index";
 import { service } from "@internal/habitat-harness/service/impl";
-import type { StructuralCheckService } from "@internal/habitat-harness/service/model/check/policy/structural/index";
+import {
+  checkCommandContext,
+  type StructuralCheckService,
+  verifyCheckSummary,
+} from "@internal/habitat-harness/service/model/check/policy/structural/index";
+import { ORPCError } from "@orpc/server";
+import { Clock, Effect } from "effect";
+import {
+  createVerifyReceipt,
+  readVerifyTargetPlan,
+  type VerifyBaseResolution,
+} from "./model/index.js";
 
 export interface VerifyModuleContext {
+  readonly checkCommandContext: typeof checkCommandContext;
+  readonly createVerifyReceipt: typeof createVerifyReceipt;
+  readonly currentTimeMillis: typeof Clock.currentTimeMillis;
   readonly epochMillisToIsoString: typeof epochMillisToIsoString;
-  readonly git: GitProviderService;
-  readonly graphite: GraphiteProviderService;
-  readonly nx: NxProviderService;
-  readonly repoRoot: string;
+  readonly observeGitStatus: ReturnType<typeof makeObserveGitStatus>;
+  readonly readVerifyTargetPlan: typeof readVerifyTargetPlanEffect;
+  readonly resolveVerifyBase: ReturnType<typeof makeResolveVerifyBase>;
+  readonly runAffectedVerification: ReturnType<typeof makeRunAffectedVerification>;
   readonly structuralCheck: StructuralCheckService;
+  readonly verifyCheckSummary: typeof verifyCheckSummary;
 }
 
-export const module = service.verify.use(({ context, next }) =>
-  next({
+export const module = service.verify.use(({ context, next }) => {
+  const resolveVerifyBase = makeResolveVerifyBase({
+    git: context.deps.git,
+    graphite: context.deps.graphite,
+    repoRoot: context.deps.repoRoot,
+  });
+  const runAffectedVerification = makeRunAffectedVerification(context.deps.nx);
+  const observeGitStatus = makeObserveGitStatus({
+    git: context.deps.git,
+    repoRoot: context.deps.repoRoot,
+  });
+  return next({
     context: {
+      checkCommandContext,
+      createVerifyReceipt,
+      currentTimeMillis: Clock.currentTimeMillis,
       epochMillisToIsoString: context.deps.epochMillisToIsoString,
-      git: context.deps.git,
-      graphite: context.deps.graphite,
-      nx: context.deps.nx,
-      repoRoot: context.deps.repoRoot,
+      observeGitStatus,
+      readVerifyTargetPlan: readVerifyTargetPlanEffect,
+      resolveVerifyBase,
+      runAffectedVerification,
       structuralCheck: context.deps.structuralCheck,
+      verifyCheckSummary,
     } satisfies VerifyModuleContext,
-  })
-);
+  });
+});
+
+function readVerifyTargetPlanEffect() {
+  return Effect.promise(() => readVerifyTargetPlan());
+}
+
+function makeRunAffectedVerification(nx: NxProviderService) {
+  return (base: string, targets: readonly string[]) =>
+    nx
+      .affected({ base, targets })
+      .pipe(Effect.map(spawnResultFromCommandResult), Effect.mapError(verifyServiceInternalError));
+}
+
+function makeObserveGitStatus(input: {
+  readonly git: GitProviderService;
+  readonly repoRoot: string;
+}) {
+  return () =>
+    input.git
+      .statusShortBranch({ cwd: input.repoRoot })
+      .pipe(Effect.map(spawnResultFromCommandResult), Effect.mapError(verifyServiceInternalError));
+}
+
+function makeResolveVerifyBase(context: {
+  readonly git: GitProviderService;
+  readonly graphite: GraphiteProviderService;
+  readonly repoRoot: string;
+}) {
+  return (base?: string) => {
+    if (base) {
+      return Effect.succeed({
+        kind: "resolved",
+        base,
+        source: "flag",
+      } satisfies VerifyBaseResolution);
+    }
+    return Effect.gen(function* () {
+      const graphiteParent = yield* context.graphite.parent({ cwd: context.repoRoot });
+      if (graphiteParent) {
+        return {
+          kind: "resolved",
+          base: graphiteParent,
+          source: "graphite-parent",
+        } satisfies VerifyBaseResolution;
+      }
+
+      const defaultBranch = yield* context.git.remoteDefaultBranch({ cwd: context.repoRoot });
+      const resolved = defaultBranch
+        ? yield* context.git.mergeBase(defaultBranch, { cwd: context.repoRoot })
+        : null;
+      if (resolved) {
+        return {
+          kind: "resolved",
+          base: resolved,
+          source: "merge-base",
+        } satisfies VerifyBaseResolution;
+      }
+      return {
+        kind: "refused",
+        message:
+          "could not resolve verify base from the remote default branch; pass --base explicitly.",
+      } satisfies VerifyBaseResolution;
+    });
+  };
+}
+
+function verifyServiceInternalError() {
+  return new ORPCError("INTERNAL_SERVER_ERROR", {
+    message: "Habitat verify service failed.",
+  });
+}
